More refactoring

+++ app/uk/gov/hmrc/thirdpartyapplication/controllers/ApplicationController.scala
@@ -307,17 +307,50 @@ class ApplicationController @Inject()(val applicationService: ApplicationService
   }
 
   def deleteApplication(id: UUID) = (Action andThen strideAuthRefiner()).async(parse.json) { implicit request: OptionalStrideAuthRequest[JsValue] => {
-    def audit(app: ApplicationData): Future[AuditResult] = {
-            Logger.info(s"Delete application ${app.id} - ${app.name}")
-            successful(uk.gov.hmrc.play.audit.http.connector.AuditResult.Success)
+      def audit(app: ApplicationData): Future[AuditResult] = {
+        // TODO Audit me
+        Logger.info(s"Delete application ${app.id} - ${app.name}")
+        successful(uk.gov.hmrc.play.audit.http.connector.AuditResult.Success)
+      }
+
+    def strideAuthenticatedApplicationDelete(deleteApplicationPayload: DeleteApplicationRequest) = {
+      // This is audited in the GK FE
+      gatekeeperService.deleteApplication(id, deleteApplicationPayload).map(_ => NoContent)
+    }
+
+    def nonStrideAuthenticatedApplicationDelete(deleteApplicationPayload: DeleteApplicationRequest): Future[Result] = {
+        val notAllowed: Result = Results.BadRequest("Cannot delete this application")
+
+        if (authConfig.canDeleteApplications) {
+          applicationService.fetch(id) flatMap {
+            case Some(app) if app.access.accessType == AccessType.STANDARD =>
+              applicationService.deleteApplication(id, deleteApplicationPayload, audit).map(_ => NoContent)
+            case _ => Future.successful(notAllowed)
+          }
+        } else {
+          Future.successful(notAllowed)
+        }
     }
 
     withJsonBody[DeleteApplicationRequest] { deleteApplicationPayload =>
-          applicationService.deleteApplication(id, deleteApplicationPayload, audit).map(_ => NoContent)
-      } recover recovery
+        if (request.isStrideAuth) {
+           strideAuthenticatedApplicationDelete(deleteApplicationPayload)
+        } else {
+          nonStrideAuthenticatedApplicationDelete(deleteApplicationPayload)
+        }
 
+      } recover recovery
     }
   }
+
+//  def deleteApplication2(id: UUID) = requiresAuthentication().async(parse.json) {
+//    implicit request =>
+//      withJsonBody[DeleteApplicationRequest] { deleteApplicationPayload =>
+//        gatekeeperService.deleteApplication(id, deleteApplicationPayload).map(_ => NoContent)
+//      } recover recovery
+//  }
 }
+
+
 
 case class ApplicationControllerConfig(fetchApplicationTtlInSecs: Int, fetchSubscriptionTtlInSecs: Int)
